+++ src/pages/RegisterPage/RegisterDesktopBackImage.jsx
@@ -1,10 +1,10 @@
 /* import assets */
-import registerImage from "../../../assets/images/register-main.png";
+import registerImage from "../../assets/images/register-main.png";
 
 /* UI */
 const RegisterDesktopBackImage = () => {
   return (
-    <>
+    <div className="relative max-h-[1000px] hidden md:block md:w-1/2">
       {/* 오버레이 그라데이션 */}
       <div
         className="absolute inset-0 z-10"
@@ -20,7 +20,7 @@ const RegisterDesktopBackImage = () => {
           backgroundImage: `url("${registerImage}")`,
         }}
       />
-    </>
+    </div>
   )
 }
