+++ src/pages/RegisterPage/RegisterForm/AgreementCheckbox.jsx
