+++ src/pages/RegisterPage/RegisterTitle.jsx
@@ -1,4 +1,4 @@
-import logo from "../../../assets/images/logo.png";
+import logo from "../../assets/images/logo.png";
 
 const RegisterTitle = () => {
   return (
