+++ src/pages/RegisterPage/loginLink.jsx
@@ -1,0 +1,12 @@
+const loginLink = () => {
+  return (
+    <p className="mb-4 text-center text-sm text-gray-600">
+      이미 계정이 있다면?{" "}
+      <a href="login" className="text-indigo-500 hover:underline">
+        Login
+      </a>
+    </p>
+  )
+}
+
+export default loginLink
